backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
bio==1.7.1
biopython==1.85
biothings_client==0.4.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
DateTime==5.5
Django==5.1.6
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.9
exceptiongroup==1.2.2
fastapi==0.115.8
Flask==3.0.3
gprofiler-official==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mygene==3.2.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
pooch==1.8.2
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.31.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
views==0.3
Werkzeug==3.1.3
XlsxWriter==3.2.2
zope.interface==7.2
django-extensions==3.2.3